scripts/requirements.txt
casparser==0.7.4
pymongo==4.6.3
watchdog==4.0.0
python-dotenv==1.0.1